+++ src/main/java/prisma/home/phe/domain/Liquidation.java
@@ -1,5 +1,6 @@
 package prisma.home.phe.domain;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 import lombok.AllArgsConstructor;
@@ -19,12 +20,12 @@ public class Liquidation {
 
   private String brand;
 
-  private Double grossPay;
+  private BigDecimal grossPay;
 
-  private Double fee;
+  private BigDecimal fee;
 
-  private Double financialCost;
+  private BigDecimal financialCost;
 
-  private Double netPay;
+  private BigDecimal netPay;
 
 }
